JAMES-2993 Create a Preview.Factory
